requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
diffusers>=0.25.0
transformers>=4.51.3
torch>=2.0.0
torchvision>=0.15.0
accelerate>=0.24.0
Pillow>=10.0.0
python-multipart>=0.0.6
pydantic>=2.0.0
numpy>=1.24.0
runpod>=1.6.0
python-jose[cryptography]>=3.3.0
aiofiles>=23.0.0